chore: split DedupStage into its own file
